created static class StreamAnalyzer
